requirements.txt
# requirements.txt
biopython>=1.79
numpy>=1.21.0
requests>=2.26.0
beautifulsoup4>=4.9.3